:tada: First TypeScript adaptation!
