+++ tsconfig.json
@@ -3,9 +3,10 @@
       "outDir": "./dist/",
       "sourceMap": true,
       "noImplicitAny": true,
-      "module": "es6",
+      "module": "CommonJS",
       "target": "es5",
-      "jsx": "react",
-      "allowJs": true
+      "allowJs": true,
+      "allowSyntheticDefaultImports" : true,
+      "esModuleInterop" : true
     }
   }
